[#49]refactor: 클래스룸 Account 초기화 기능 classroomAccount 도메인으로 이동, ActionStatus.DEFAULT 추가
