+++ src/main/java/soma/edupimeta/classroom/ClassroomController.java
@@ -11,7 +11,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import soma.edupimeta.classroom.account.models.ActionInitRequest;
 import soma.edupimeta.classroom.models.ClassroomInfoResponse;
 import soma.edupimeta.classroom.models.CreateClassroomRequest;
 import soma.edupimeta.classroom.models.MyClassroomResponse;
@@ -53,18 +52,6 @@ public class ClassroomController implements ClassroomOpenApi {
         return ResponseEntity
             .status(HttpStatus.OK)
             .build();
-    }
-
-    @PostMapping("/v1/classroom/action/init")
-    public ResponseEntity<Long> initActionStatusBy(@RequestBody ActionInitRequest actionInitRequest) {
-        log.info("Init classroom id: {}", actionInitRequest.getClassroomId());
-        long updateCount = classroomService.initAllActionStatusBy(
-            actionInitRequest.getClassroomId()
-        );
-
-        return ResponseEntity
-            .status(HttpStatus.OK)
-            .body(updateCount);
     }
 
     @Override
