+++ src/main/java/soma/edupimeta/classroom/account/service/domain/ActionStatus.java
@@ -8,7 +8,7 @@ import soma.edupimeta.classroom.account.exception.ClassroomAccountException;
 
 @Getter
 public enum ActionStatus {
-    ING(0), HELP(1), COMPLETE(2);
+    DEFAULT(0), ING(1), HELP(2), COMPLETE(3);
 
     private final int value;
